+++ src/OpenTelemetry.Instrumentation.Hangfire/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+* Update `OpenTelemetry.Api.ProviderBuilderExtensions` to `1.7.0`.
+  * Update `OpenTelemetry.Api` to `1.7.0`.
+  ([#1508](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/1508))
+
 ## 1.6.0-beta.1
 
 Released 2023-Dec-20